*** empty log message ***
git-svn-id: https://svn.library.emory.edu/svn/slavevoyages/trunk@636 2440b244-9809-4ec7-bebd-234906d9abf5

+++ src/edu/emory/library/tast/reditor/FieldValueLong.java
@@ -45,7 +45,7 @@ public class FieldValueLong extends FieldValueText
 		return valid;
 	}
 	
-	public Long getInteger()
+	public Long getLong()
 	{
 		if (!isValid())
 		{
@@ -76,6 +76,5 @@ public class FieldValueLong extends FieldValueText
 			setValue("");
 		}
 	}
-	
 
 }
